+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#130] Add possibility to add custom annotations for `LoadBalancer` service.
 
 ## [v4.1.0] - 2025-07-16
 ### Fixed